feat: update

+++ service/user/data/request/doc/doc.go
@@ -9,7 +9,7 @@ type CreateDocRequest struct {
 	Title   string `json:"title" binding:"min=1,max=250"` // 标题
 	Content string `json:"content" binding:""`            // 文档内容
 	Pid     int64  `json:"pid" binding:""`
-	GroupId int    `json:"groupId" binding:""`
+	GroupId int64  `json:"groupId" binding:""`
 	IsTop   int    `json:"isTop" binding:""` // 是否置顶
 }
